FIX QuizList, ListeningQuiz level, quizType, musicId set Unique 3

+++ backend/src/main/java/com/example/melLearnBE/dto/model/MusicDto.java
@@ -21,8 +21,8 @@ public class MusicDto {
     private Long id;
     private String musicId;
     private Language language;
-    private Long liked;
-    private Long disLike;
+    private int liked;
+    private int disLike;
     private boolean speaking;
     private boolean grammar;
     private boolean listening;
